fix: missing give insert payment keys
